Remove dict related functions from summary_services
